requirements.txt
matplotlib
numpy
pandas
Pillow
scikit_learn
seaborn
torch
tqdm
transformers
datasets
accelerate
tensorboard